newsecurityy/backend/requirements.txt
Django==5.0.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
